I added an extra line at bottom of each file.
I changed the names of T564 and T565 to T583 and T584.

I changed T584's hypothesis to include (P000137: false).

I fixed the name field for Topology (Munkres) in T583 and T584.

I added Willard as a reference to T583

I added Wikipedia as a reference to P199.

I added two additional equivalent definitions to P199.

I added a proof of P199.

I added a link to a mathse answer for a proof of P199.